rag-with-access-control/backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
langchain-google-genai
chromadb
langchain-community
PyJWT
pytest
httpx